pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "budgeteer"
version = "0.1.0"
description = "Interactive CLI for recording expenses"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "questionary>=2.0",
  "rich>=13.7",
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "platformdirs>=4.2",
]

[project.scripts]
budgeteer = "budgeteer.cli:app"

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "ruff>=0.5",
]

[tool.hatch.build.targets.wheel]
packages = ["src/budgeteer"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
